// types/schema-types.ts
type Validation = (rule: RuleType) => RuleType;

type HiddenField = (({ parent, document, value }: any) => boolean) | boolean;

export interface CommonFieldProps {
  type: string;
  name: string;
  title?: string;
  description?: string;
  fieldset?: string;
  hidden?: HiddenField;
  inputComponent?: React.ReactNode;
  readOnly?: boolean;
  validation?: Validation;
  preview?: Preview;
  group?: string;
  icon?: ReactComponentLike; // is only available for elements of which include a block
  options?: {
    [key: string]: any;
  };
}

export interface StringField extends CommonFieldProps {
  options?: {
    list: { title: string; value: string }[];
    layout?: string;
  };
}

interface TextField extends CommonFieldProps {
  rows: number;
}

export interface Span {
  _type: "span";
  text: string;
}

export interface BlockField extends CommonFieldProps {
  _type: "block";
  styles: {
    title: string;
    value: string;
    blockEditor?: {
      render: ReactComponentLike | JSX.Element;
    };
    icon?: ReactComponentLike;
  }[];
  marks?: {
    decorators?: { title: string; value: string }[];
    annotations?: (ObjectField & {
      blockEditor?: { icon: React.ReactNode } | React.ReactNode;
    })[];
  };
  lists?: { title: string; value: string }[];
  children: (Field | Span)[];
}

type ArrayOf =
  | ObjectField
  | ReferenceField
  | ImageField
  | { type: string }
  | BlockField;

export interface ArrayField extends CommonFieldProps {
  type: "array";
  name: string;
  of: ArrayOf[];
}

type FilterFunctionResult = { filter: string; filterParams?: string };
type ReferenceType = "post" | "route" | "legal" | "page";
type FilterFunction = (args: {
  document: { [key: string]: any };
  parentPath: string[];
  parent: Record<string, unknown>[];
}) => FilterFunctionResult;

export interface ReferenceField extends CommonFieldProps {
  to: { type: ReferenceType }[];
  options?: {
    filter: string | FilterFunction;
    filterParams?: { [key: string]: string };
  };
}

export interface ImageField extends CommonFieldProps {
  type: "image";
  options?: {
    hotspot?: boolean;
  };
  fields?: Field[];
}

export type Field =
  | CommonFieldProps
  | StringField
  | TextField
  | ArrayField
  | ReferenceField
  | ImageField
  | ObjectField
  | BlockField;

interface Preview {
  select?: { [key: string]: string };
  prepare?: (selection: { [key: string]: any }) => {
    title?: string;
    subtitle?: string;
    media?: string | React.ReactNode;
  };
  component?: (props: PreviewProps) => React.ReactElement;
}

interface Fieldset {
  name: string;
  title: string;
  options?: { collapsible: boolean; collapsed?: boolean };
}

export interface ObjectField extends CommonFieldProps {
  type: "object";
  fields: Field[];
  fieldsets?: Fieldset[];
  initialValue?: { [key: string]: any };
  options?: { collapsible?: boolean; collapsed?: boolean };
}